requirements.txt
annotated-types==0.6.0
ansicon==1.89.0
anyio==4.1.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.1.0
binaryornot==0.4.4
black==24.10.0
blessed==1.20.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cookiecutter==2.5.0
cramjam==2.8.4
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==4.2.13
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-yasg==1.21.7
editor==1.6.6
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
Flask==3.0.0
fonttools==4.45.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.4
inflection==0.5.1
inquirer==3.3.0
ipykernel==6.27.0
ipython==8.17.2
ipyvue==1.10.1
ipyvuetify==1.8.10
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jinxed==1.2.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lxml==5.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
Mesa==2.1.4
Mesa-Viz-Tornado==0.1.3
mypy-extensions==1.0.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
odfpy==1.4.1
openai==1.28.1
openpyxl==3.1.5
packaging==23.2
pandas==2.1.3
parsimonious==0.10.0
parso==0.8.3
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.0.0
portion==2.5.0
progressbar2==4.5.0
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2-binary==2.9.10
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.9.0
pymdown-extensions==10.4
pymongo==4.10.1
pyparsing==3.1.1
PyPDF2==3.0.1
pysd==3.14.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.1
python-snappy==0.7.3
python-utils==3.9.0
pytz==2023.3.post1
pyxlsb==1.0.10
PyYAML==6.0.1
pyzmq==25.1.1
reacton==1.8.0
readchar==4.1.0
referencing==0.31.0
regex==2024.9.11
requests==2.31.0
rest-framework-simplejwt==0.0.2
rich==13.7.0
rich-click==1.7.1
rpds-py==0.13.1
runs==1.2.2
six==1.16.0
sniffio==1.3.0
solara==1.23.0
sortedcontainers==2.4.0
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.32.0.post1
text-unidecode==1.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
virtualenv==20.26.6
virtualenvwrapper-win==1.2.7
watchdog==3.0.0
wcwidth==0.2.12
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
xlrd==2.0.1
XlsxWriter==3.2.0
xmod==1.8.1
